scraper/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
anthropic==0.49.0
sentence-transformers==3.4.1
boto3==1.37.19
